requirements.txt
wget
tensorflow
opencv-python
Pillow
numpy
scikit-learn
mediapipe
